Split code into multiple files, try create FanController
