+++ Cargo.toml
@@ -15,3 +15,4 @@ linux-embedded-hal = "0.3.2"
 machine-ip = "0.2.1"
 profont = "0.7.0"
 display-interface = "0.4.1"
+i2cdev = "0.6.0"
